docs: add disclaimer
There've been more people (beginners, mainly) using my dotfiles -- some
even use it *verbatim*, shockingly. They then call on me as if I were an
expert in the NixOS space. I've tried to make my docs clearer in 809ee25
and 3070ee6 for the stubborn, but if that doesn't help, there are better
sources of help elsewhere, mentioned in my disclaimer.

Ref 809ee25f738b