refactor!: remove pkgs.unstable
BREAKING CHANGE: This removes the pkgs.unstable.* overlays and
nixpkgs-unstable input. At first, I used this to have a separate channel
for faster moving packages, but it's proven to be a maintenance burden.
I follow nixpkgs-unstable closely enough anyway, and can pull nixpkgs
tarballs directly when I need a specific version of something in the
future.