refactor: remove disko refs
Never did use it, and am too nervous to rely on it on workstations,
where a wrong command could result in data loss. Redeployments are so
rare for my workstations, I *will* forget them, and breaking changes
from upstream could catch me by surprise at the worst time and in the
worst way.